Convert Morgoth to Kapacitor UDF (#29)
* start refactor for Kapacitor UDF

* working UDF

* update to use minSupport

* output anomalous batch, jsdiv flexible bounds

* add explicit log handling

* add socket support

* add docs about socket UDF

* update docs ref

+++ counter/counter.go
@@ -1,9 +1,9 @@
 package counter
 
 type Counter interface {
-	// Count a fingerprint and return the number of times
-	// that fingerprint has been seen
-	Count(Countable) int
+	// Count a fingerprint and return the support for the item.
+	// support = count / total
+	Count(Countable) float64
 }
 
 type Countable interface {
